Rename datetime module to times
