+++ .docs/source/refs/index.rst
@@ -4,7 +4,7 @@
 
       common
       containers
-      datetime
       debug
       profiler
       strings
+      times
