+++ pure_utils/__init__.py
@@ -4,7 +4,7 @@ __version__ = "0.6.0"
 
 from .common import *  # noqa: F401, F403
 from .containers import *  # noqa: F401, F403
-from .datetime import *  # noqa: F401, F403
 from .debug import *  # noqa: F401, F403
 from .profiler import *  # noqa: F401, F403
 from .strings import *  # noqa: F401, F403
+from .times import *  # noqa: F401, F403
